Organized components into folders & added README
