+++ src/components/Layout/Layout.sass
@@ -74,7 +74,7 @@ div[role="dialog"]
 .App
     overflow: hidden
 
-    background-image: url('../images/bg.webp')
+    background-image: url('../../images/bg.webp')
     background-color: #030c15
     background-size: 300px
     min-height: 100vh
