+++ src/components/Paginator/Paginator.sass
